moving django dependencies apart from the rest of the extra fields
